auraflow/requirements.txt
openai==1.35.3
stripe==10.1.0
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
praw==7.7.1 # For Reddit scraping
pytest==8.2.2
pytest-mock==3.14.0
